LayoutTests/imported/w3c/web-platform-tests/cookie-store/encoding.https.any-expected.txt


FAIL BOM not stripped from name promise_test: Unhandled rejection with value: object "TypeError: null is not an object (evaluating 'cookie.name')"
FAIL BOM not stripped from value assert_equals: expected "﻿value" but got "ï»¿value"
